🎉 1.0.8 Enhanced themes and exciting new features that you'ill love!
The Shiba Inu theme keeps getting better and better, and I'm happy to say that my CSS skills are improving along with it. With each update of the Shiba Inu theme, I am able to improve my CSS skills. I'm proud to say that the latest version is more beautiful than ever, and I hope to make the next one even better