Welcome to Shiba Inu theme! 🐕
I regret to inform you that I, Farouk, the founder of the Shiba Inu Theme, have left Obsidian in March 2023 due to personal reasons. As a result, it's uncertain whether any future updates to the theme will be made. I apologize for any confusion caused by my departure and I want to express my gratitude to the Obsidian community for their support and appreciation for the Shiba Inu Theme. Creating and sharing this theme with you all has been a wonderful experience.

*Despite my efforts to contribute to the Obsidian community, I found that the app wasn't meeting my personal needs and goals.*